requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1 #stable with passlib
python-dotenv
email-validator
pydantic[email]
slowapi
limits
